requirements.txt
aniso8601==3.0.2
click==6.7
dnspython==1.16.0
eventlet==0.25.1
Flask==1.0.2
Flask-Compress==1.4.0
Flask-HTTPAuth==3.2.4
Flask-RESTful==0.3.6
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
Pillow==6.2.0
PyMySQL==0.9.3
pysodium==0.7.0.post0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
Werkzeug==0.15.5
